refactor(ai): extract treasury budget pin cases under residual fat-pin gate
Split treasury_planner_treasury_budget_test into topical *_cases modules
plus shared support, and gate the thin contract in repo.ai_residual_fat_pin_cases_required.
