+++ SPEC/program/repo-lint.md
@@ -99,7 +99,7 @@
 | `tool/check_ai_colonial_phase_dispatch_test_shared_fixtures.dart` | Forbid local `_colonialLiteGame` / `_colonialGame` redeclarations in `packages/colonizethis_ai/test/planning/phase_planner_dispatch_test.dart`. Pins MUST import `buildPhasePlannerDispatchColonialLiteGame` / `buildPhasePlannerDispatchColonialGame` from `packages/colonizethis_ai/test/support/colonial_phase_planner_test_support.dart`. The gate also fails when that shared support file is missing (Refs #3977); rule `repo.ai_colonial_phase_test_shared_fixtures` |
 | `tool/check_ai_colonial_military_naval_pin_suite_size.dart` | Forbid oversize COLONIAL military / naval / lite-naval contracts: `colonial_phase_planner_military_test.dart`, `colonial_phase_planner_naval_test.dart`, and `colonial_phase_planner_colonial_lite_naval_test.dart` with more than **750** physical lines that do not import a sibling `*_cases.dart` are violations (Refs #3977, #3997); rule `repo.ai_colonial_military_naval_pin_cases_required` |
 | `tool/check_ai_planning_cases_suite_size.dart` | Forbid oversize AI planning case libraries: under `packages/colonizethis_ai/test/planning/`, any `*_cases.dart` with more than **650** physical lines that is not registered in `aiPlanningCasesSuiteSizeAllowlist` (path → issue Ref) is a violation. Prefer topic-splits over allowlist entries (Refs #3997); rule `repo.ai_planning_cases_suite_size` |
-| `tool/check_ai_residual_fat_pin_suite_size.dart` | Forbid oversize Phase-8 residual fat-pin contracts: gated basenames under `packages/colonizethis_ai/test/planning/` (currently `observer_goal_phase_test.dart`) with more than **750** physical lines that do not import a sibling `*_cases.dart` are violations (Refs #3997); rule `repo.ai_residual_fat_pin_cases_required` |
+| `tool/check_ai_residual_fat_pin_suite_size.dart` | Forbid oversize Phase-8 residual fat-pin contracts: gated basenames under `packages/colonizethis_ai/test/planning/` (currently `observer_goal_phase_test.dart`, `treasury_planner_treasury_budget_test.dart`) with more than **750** physical lines that do not import a sibling `*_cases.dart` are violations (Refs #3997); rule `repo.ai_residual_fat_pin_cases_required` |
 | `tool/check_ai_economy_regiment_build_test_shared_fixtures.dart` | Forbid local economy regiment-build Game factory redeclarations in `packages/colonizethis_ai/test/planning/economy_planner_regiment_build*_test.dart` and `*_cases.dart` (the shared `*_support.dart` is exempt). A violation is any `Game _regimentRebuildProductionGame` / `_castIronImprovementInputGame` / `_supplierCastIronSourceGame` / `_castIronFeedstockCoavailabilityGame` / `_castIronStagingNoFabricGateGame` / `_castIronLabourPeasantRecruitFabricStagingGame` declaration inside those pins. Pins MUST import the public factories from `packages/colonizethis_ai/test/planning/economy_planner_regiment_build_input_support.dart`. The gate also fails when that shared support file is missing (Refs #3972); rule `repo.ai_economy_regiment_build_test_shared_fixtures` |
 | `tool/check_ai_test_no_matrix_part_shards.dart` | Forbid planning matrix part shards that have already grown a sibling support library. Under `packages/colonizethis_ai/test/planning/`, any `*_partN_test.dart` with `N ≥ 2` is a violation when a same-directory `{stem}_support.dart` exists for that stem (e.g. `foo_matrix_part2_test.dart` next to `foo_matrix_support.dart`). Suites mid-migration without a support file are not flagged. Consolidated suites keep one `*_test.dart` contract plus support / case-table libraries (Refs #3941); rule `repo.ai_test_no_matrix_part_shards` |
 | `tool/check_ai_test_mirrors_lib.dart` | Forbid `*_test.dart` sources directly at the `packages/colonizethis_ai/test/` root: AI production code lives under `lib/src/{perception,planning,social,tactical,util}/`, so the testing rule (`colonizethis-testing.mdc`) requires `test/` to mirror `lib/`. Every AI unit test must live in a mirroring subdirectory such as `test/planning/` rather than flat in `test/`. A `*_test.dart` is flagged when the path remainder after the `packages/colonizethis_ai/test/` prefix contains no further separator; nested files such as `test/planning/war_desire_score_test.dart` are compliant. Non-`*_test.dart` scaffolding helpers at the root are out of scope here (governed by `repo.ai_test_no_duplicate_scaffolding`). A documented allowlist (`aiTestMirrorsLibAllowlist`) permits package-level surface tests (public barrel / DI exports: `ai_config_test.dart`, `ai_di_export_test.dart`, `seed_bundle_test.dart`) and cross-cutting integration/diagnostic families (`seed42_*`, `observer_*`, `full_ai_*`) that drive whole-AI/observer runs rather than mirroring a single `lib/src/` source. Mirrors the sibling layout gates `repo.orders_test_mirrors_lib` / `repo.combat_test_mirrors_lib` (Refs #3717); rule `repo.ai_test_mirrors_lib` |
@@ -381,7 +381,7 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a gated military, naval, or lite-naval contract file has more than 750 physical lines and does not import a `*_cases.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_colonial_military_naval_pin_cases_required`, then the run fails listing the offending file and exits `1` (Refs #3977, #3997).
 - Given every `*_cases.dart` under `packages/colonizethis_ai/test/planning/` either has ≤650 physical lines or is registered in `aiPlanningCasesSuiteSizeAllowlist` with an issue Ref, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_planning_cases_suite_size`, then the rule passes without violations and exits `0` (Refs #3997).
 - Given a `packages/colonizethis_ai/test/planning/*_cases.dart` file has more than 650 physical lines and is not allowlisted, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_planning_cases_suite_size`, then the run fails listing the offending file and exits `1` (Refs #3997).
-- Given each gated Phase-8 residual fat-pin contract (`observer_goal_phase_test.dart`) either has ≤750 physical lines or imports a sibling `*_cases.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_residual_fat_pin_cases_required`, then the rule passes without violations and exits `0` (Refs #3997).
+- Given each gated Phase-8 residual fat-pin contract (`observer_goal_phase_test.dart`, `treasury_planner_treasury_budget_test.dart`) either has ≤750 physical lines or imports a sibling `*_cases.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_residual_fat_pin_cases_required`, then the rule passes without violations and exits `0` (Refs #3997).
 - Given a gated residual fat-pin contract file has more than 750 physical lines and does not import a `*_cases.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_residual_fat_pin_cases_required`, then the run fails listing the offending file and exits `1` (Refs #3997).
 - Given `packages/colonizethis_ai/test/planning/economy_planner_regiment_build*_test.dart` pins import shared Game factories from `test/planning/economy_planner_regiment_build_input_support.dart` and do not redeclare `_regimentRebuildProductionGame` / `_castIronImprovementInputGame` / `_supplierCastIronSourceGame` / `_castIronFeedstockCoavailabilityGame` / `_castIronStagingNoFabricGateGame` / `_castIronLabourPeasantRecruitFabricStagingGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_economy_regiment_build_test_shared_fixtures`, then the rule passes without violations and exits `0` (Refs #3972).
 - Given an `economy_planner_regiment_build*_test.dart` file that redeclares `Game _regimentRebuildProductionGame` or `Game _castIronImprovementInputGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_economy_regiment_build_test_shared_fixtures`, then the run fails listing each offending file and exits `1` (Refs #3972).
