+++ test/check_ai_residual_fat_pin_suite_size_test.dart
@@ -7,12 +7,12 @@ import '../tool/check_ai_residual_fat_pin_suite_size.dart';
 
 void main() {
   group('runCheckAiResidualFatPinSuiteSize', () {
-    test('fails when observer_goal_phase pin is oversize without cases', () {
+    test('fails when gated residual fat pin is oversize without cases', () {
       final temp = Directory.systemTemp.createTempSync('ai-residual-fat-');
       try {
         _writePlanningTest(
           temp,
-          'observer_goal_phase_test.dart',
+          'treasury_planner_treasury_budget_test.dart',
           '${List.filled(800, '// pad').join('\n')}\nvoid main() {}\n',
         );
         final errors = <String>[];
@@ -22,20 +22,23 @@ void main() {
           err: errors.add,
         );
         expect(exitCode, 1);
-        expect(errors.join('\n'), contains('observer_goal_phase_test.dart'));
+        expect(
+          errors.join('\n'),
+          contains('treasury_planner_treasury_budget_test.dart'),
+        );
         expect(errors.join('\n'), contains('*_cases.dart'));
       } finally {
         temp.deleteSync(recursive: true);
       }
     });
 
-    test('passes when observer_goal_phase pin imports *_cases.dart', () {
+    test('passes when gated residual fat pin imports *_cases.dart', () {
       final temp = Directory.systemTemp.createTempSync('ai-residual-fat-ok-');
       try {
         _writePlanningTest(
           temp,
-          'observer_goal_phase_test.dart',
-          "import 'observer_goal_phase_phase_and_declare_war_cases.dart';\n"
+          'treasury_planner_treasury_budget_test.dart',
+          "import 'treasury_planner_treasury_budget_deficit_clamp_cases.dart';\n"
           '${List.filled(800, '// pad').join('\n')}\n'
           'void main() {}\n',
         );
